Removed Curve ExtendedMethods.
