+++ src/GShark/Geometry/NurbsCurve.cs
@@ -1,5 +1,6 @@
 ﻿#nullable enable
 using GShark.Core;
+using GShark.Interfaces;
 using GShark.Operation;
 using GShark.Operation.Utilities;
 using System;
@@ -7,7 +8,7 @@ using System.Collections.Generic;
 using System.Globalization;
 using System.Linq;
 using System.Text;
-using GShark.Interfaces;
+using GShark.ExtendedMethods;
 
 namespace GShark.Geometry
 {
@@ -142,7 +143,7 @@ namespace GShark.Geometry
         /// <returns>True if the curve is closed.</returns>
         public bool IsClosed()
         {
-            Point3 pt0 = Evaluation.CurvePointAt(this,0.0);
+            Point3 pt0 = Evaluation.CurvePointAt(this, 0.0);
             Point3 pt1 = Evaluation.CurvePointAt(this, 1.0);
             return pt0.EpsilonEquals(pt1, GSharkMath.Epsilon);
         }
@@ -310,7 +311,7 @@ namespace GShark.Geometry
 
             List<Vector3> derivatives = Evaluation.RationalCurveDerivatives(this, t, 2);
 
-            Vector3 normal = (derivatives[2].Length == 0.0) 
+            Vector3 normal = (derivatives[2].Length == 0.0)
                 ? Vector3.PerpendicularTo(derivatives[1])
                 : Analyze.Curvature(derivatives[1], derivatives[2]);
 
@@ -347,7 +348,6 @@ namespace GShark.Geometry
         /// Computes the parameter along the curve which coincides with a given length.
         /// </summary>
         /// <param name="segmentLength">Length of segment to measure. Must be less than or equal to the length of the curve.</param>
-        /// <param name="tolerance">If set less or equal 0.0, the tolerance used is 1e-10.</param>
         /// <returns>The parameter on the curve at the given length.</returns>
         public double ParameterAtLength(double segmentLength)
         {
@@ -371,7 +371,7 @@ namespace GShark.Geometry
         {
             if (t <= 0.0)
             {
-               return 0.0;
+                return 0.0;
             }
 
             if (t >= 1.0)
@@ -430,6 +430,218 @@ namespace GShark.Geometry
             }
 
             return Fitting.InterpolatedCurve(offsetPts, Degree);
+        }
+
+        /// <summary>
+        /// Divides a curve for a given number of time, including the end points.<br/>
+        /// The result is not split curves but a collection of t values and lengths that can be used for splitting.<br/>
+        /// As with all arc length methods, the result is an approximation.
+        /// </summary>
+        /// <param name="numberOfSegments">The number of parts to split the curve into.</param>
+        /// <returns>A tuple define the t values where the curve is divided and the lengths between each division.</returns>
+        public (List<Point3> Points, List<double> Parameters) Divide(int numberOfSegments)
+        {
+            if (numberOfSegments < 2)
+            {
+                throw new ArgumentException("Number of segments must be greater than 1.", nameof(numberOfSegments));
+            }
+
+            var divideResult = Operation.Divide.CurveByCount(this, numberOfSegments);
+            var points = divideResult.Select(PointAt).ToList();
+            return (points, divideResult);
+        }
+
+        /// <summary>
+        /// Divides a curve for a given max segment length, including the end points.<br/>
+        /// The result is not split curves but a collection of t values and lengths that can be used for splitting.<br/>
+        /// As with all arc length methods, the result is an approximation.
+        /// </summary>
+        /// <param name="maxSegmentLength">The maximum length the segments have to be split in.</param>
+        /// <param name="equalSegmentLengths">Force to have all the segments of the same lengths.</param>
+        /// <returns>A tuple define the t values where the curve is divided and the lengths between each division.</returns>
+        public (List<Point3> Points, List<double> Parameters) Divide(double maxSegmentLength, bool equalSegmentLengths = false)
+        {
+            if (maxSegmentLength <= 0)
+            {
+                throw new ArgumentException("Segment length must be greater than 0.", nameof(maxSegmentLength));
+            }
+
+            var len = maxSegmentLength;
+            if (equalSegmentLengths)
+            {
+                List<NurbsCurve> curves = Modify.DecomposeCurveIntoBeziers(this);
+                List<double> curveLengths = curves.Select(nurbsCurve => Analyze.BezierCurveLength(nurbsCurve)).ToList();
+                double totalLength = curveLengths.Sum();
+
+                len = totalLength / Math.Ceiling(totalLength / maxSegmentLength);
+            }
+
+            var (tValues, lengths) = Operation.Divide.CurveByLength(this, len);
+            var points = tValues.Select(PointAt).ToList();
+
+            return (points, tValues);
+        }
+
+        /// <summary>
+        /// Creates rotation minimized perpendicular frames (RMF) at given t parameters along the curve.<br/>
+        /// Double reflection method taken from Wang, W., J¨uttler, B., Zheng, D., and Liu, Y. 2008. "Computation of rotation minimizing frame."<br/>
+        /// https://www.microsoft.com/en-us/research/wp-content/uploads/2016/12/Computation-of-rotation-minimizing-frames.pdf
+        /// </summary>
+        /// ///<param name="uValues">The curve parameter values to locate perpendicular curve frames</param>
+        public List<Plane> PerpendicularFrames(List<double> uValues)
+        {
+            var pointsOnCurve = uValues.Select(PointAt).ToList(); //get points at t values
+            var pointsOnCurveTan = uValues.Select(t => Evaluation.RationalCurveTangent(this, t)).ToList(); //get tangents at t values
+            var firstParameter = uValues[0]; //get first t value
+
+            //Create initial frame at first parameter
+            var origin = PointAt(firstParameter);
+            var crvTan = Evaluation.RationalCurveTangent(this, firstParameter);
+            var crvNormal = Vector3.PerpendicularTo(crvTan);
+            var yAxis = Vector3.CrossProduct(crvTan, crvNormal);
+            var xAxis = Vector3.CrossProduct(yAxis, crvTan);
+
+            //Set initial frame
+            Plane[] perpFrames = new Plane[pointsOnCurve.Count];
+            perpFrames[0] = new Plane(origin, xAxis, yAxis);
+
+            //Given boundary data(x0, t0; x1, t1) and an initial right-handed
+            //orthonormal frame U0 = (r0, s0, t0) at x0, the next frame U1 = (r1, s1, t1)
+            //at x1 for RMF approximation is computed by the double reflection method in
+            //the following two steps.
+            //
+            //Step 1.Let R1 denote the reflection in the bisecting plane of the points x0
+            //and x1(see Figure 4).Use R1 to map U0 to a left - handed orthonormal frame
+            //UL0 = (rL0, sL0, tL0).
+            //
+            //Step 2.Let R2 denote the reflection in the bisecting plane of the points x1 + tL
+            //0 and x1 +t1. Use R2 to map UL0 to a right - handed orthonormal frame U1 = (r1, s1, t1).
+            //Output U1.
+
+            for (int i = 0; i < pointsOnCurve.Count - 1; i++)
+            {
+                Vector3 v1 = pointsOnCurve[i + 1] - pointsOnCurve[i]; //compute reflection vector of R1
+                double c1 = v1 * v1;
+                Vector3 rLi = perpFrames[i].XAxis - (2 / c1) * (v1 * perpFrames[i].XAxis) * v1; //compute reflected rL vector by R1
+                Vector3 tLi = pointsOnCurveTan[i] - (2 / c1) * (v1 * pointsOnCurveTan[i]) * v1; //compute reflected tL vector by R1
+                Vector3 v2 = pointsOnCurveTan[i + 1] - tLi; //compute reflection vector of R2
+                double c2 = v2 * v2;
+                Vector3 rNext = rLi - (2 / c2) * (v2 * rLi) * v2; //compute reflected r vector by R2
+                var sNext = Vector3.CrossProduct(pointsOnCurveTan[i + 1], rNext); //compute vector s[i+1] of next frame
+
+                //create output frame
+                var frameNext = new Plane { Origin = pointsOnCurve[i + 1], XAxis = rNext, YAxis = sNext };
+                perpFrames[i + 1] = frameNext; //output frame
+            }
+
+            return perpFrames.ToList();
+        }
+
+        /// <summary>
+        /// Splits a curve into two parts at a given parameter.
+        /// </summary>
+        /// <param name="t">The parameter at which to split the curve.</param>
+        /// <returns>Two NurbsCurve objects.</returns>
+        public List<NurbsCurve> SplitAt(double t)
+        {
+            int degree = Degree;
+
+            List<double> knotsToInsert = CollectionHelpers.RepeatData(t, degree + 1);
+
+            NurbsCurve refinedCurve = Modify.CurveKnotRefine(this, knotsToInsert);
+
+            int s = Knots.Span(degree, t);
+
+            KnotVector knots0 = refinedCurve.Knots.ToList().GetRange(0, s + degree + 2).ToKnot();
+            KnotVector knots1 = refinedCurve.Knots.GetRange(s + 1, refinedCurve.Knots.Count - (s + 1)).ToKnot();
+
+            List<Point4> controlPoints0 = refinedCurve.ControlPoints.GetRange(0, s + 1);
+            List<Point4> controlPoints1 = refinedCurve.ControlPoints.GetRange(s + 1, refinedCurve.ControlPointLocations.Count - (s + 1));
+
+            return new List<NurbsCurve> { new NurbsCurve(degree, knots0, controlPoints0), new NurbsCurve(degree, knots1, controlPoints1) };
+        }
+
+        /// <summary>
+        /// Splits a curve at given parameters and returns the segments as curves.
+        /// </summary>
+        /// <param name="parameters">The parameters at which to split the curve. Values should be between 0.0 and 1.0.</param>
+        /// <returns>Collection of curve segments.</returns>
+        public List<NurbsCurve> SplitAt(double[] parameters)
+        {
+            var curves = new List<NurbsCurve>();
+            if (parameters.Length == 0)
+            {
+                curves.Add(this);
+                return curves;
+            }
+
+            var sortedParameters = parameters.OrderBy(x => x).ToArray();
+            Interval curveDomain = Knots.GetDomain(this.Degree);
+
+            if (Math.Abs(sortedParameters[0] - curveDomain.T0) > GSharkMath.MaxTolerance)
+            {
+                var tempParams = new double[sortedParameters.Length + 1];
+                tempParams[0] = curveDomain.T0;
+                for (var i = 0; i < sortedParameters.Length; i++)
+                {
+                    tempParams[i + 1] = sortedParameters[i];
+                }
+                sortedParameters = tempParams;
+            }
+
+            if (Math.Abs(sortedParameters[sortedParameters.Length - 1] - curveDomain.T1) > GSharkMath.MaxTolerance)
+            {
+                Array.Resize(ref sortedParameters, sortedParameters.Length + 1);
+                sortedParameters[sortedParameters.Length - 1] = curveDomain.T1;
+            }
+
+            for (int i = 0; i < sortedParameters.Length - 1; i++)
+            {
+                curves.Add(SubCurve(new Interval(sortedParameters[i], sortedParameters[i + 1])));
+            }
+
+            return curves;
+        }
+
+        /// <summary>
+        /// Extract sub-curve defined by domain.
+        /// </summary>
+        /// <param name="domain">Domain of sub-curve</param>
+        /// <returns>NurbsCurve.</returns>
+        public NurbsCurve SubCurve(Interval domain)
+        {
+            int degree = Degree;
+            int order = degree + 1;
+            Interval subCurveDomain = domain;
+
+            //NOTE: Handling decreasing domain by flipping it to maintain direction of original curve in sub-curve. Is this what we want?
+            if (domain.IsDecreasing)
+            {
+                subCurveDomain = new Interval(domain.T1, domain.T0);
+            }
+
+            var isT0AtStart = Math.Abs(subCurveDomain.T0 - Knots[0]) < GSharkMath.MaxTolerance;
+            var isT1AtEnd = Math.Abs(subCurveDomain.T1 - Knots[Knots.Count - 1]) < GSharkMath.MaxTolerance;
+
+            if (isT0AtStart && isT1AtEnd)
+            {
+                return this;
+            }
+
+            if (isT0AtStart || isT1AtEnd)
+            {
+                return isT0AtStart ? SplitAt(subCurveDomain.T1)[0] : SplitAt(subCurveDomain.T0)[1];
+            }
+
+            List<double> knotsToInsert = CollectionHelpers.RepeatData(domain.T0, order).Concat(CollectionHelpers.RepeatData(domain.T1, degree + 1)).ToList();
+            NurbsCurve refinedCurve = Modify.CurveKnotRefine(this, knotsToInsert);
+
+            var subCurveControlPoints = refinedCurve.ControlPoints.GetRange(order, order);
+            var subCrvCrtlPtsLocations = subCurveControlPoints.Select(Point4.PointDehomogenizer).ToList();
+
+            var subCurve = new NurbsCurve(subCrvCrtlPtsLocations, Degree);
+
+            return subCurve;
         }
 
         /// <summary>
